+++ packages/diagram/src/index.ts
@@ -30,10 +30,11 @@ export * from "./layout/elements/stackLayoutConfig.js";
 export * from "./layout/elements/styledElementLayoutConfig.js";
 export * from "./layout/elements/textLayoutConfig.js";
 export * from "./layout/elements/vboxLayoutConfig.js";
+export * from "./layout/engine/layout.js";
+export * from "./layout/engine/layoutEngine.js";
 export * from "./layout/diagramLayoutResult.js";
 export * from "./layout/layoutCache.js";
 export * from "./layout/layoutElement.js";
-export * from "./layout/layoutEngine.js";
 export * from "./layout/layouts.js";
 export * from "./module/base/diagramModule.js";
 export * from "./module/base/dslModule.js";
