+++ packages/diagram/src/layout/elements/canvas/canvasConnectionSegmentLayoutConfig.ts
@@ -1,10 +1,9 @@
-import { FullObject, RuntimeError, nullType } from "@hylimo/core";
+import { FullObject, nullType } from "@hylimo/core";
 import { Size } from "@hylimo/diagram-common";
 import { canvasPointType } from "../../../module/base/types.js";
 import { AttributeConfig, ContentCardinality, LayoutElement, SizeConstraints } from "../../layoutElement.js";
-import { Layout } from "../../layoutEngine.js";
+import { Layout } from "../../engine/layout.js";
 import { ElementLayoutConfig } from "../elementLayoutConfig.js";
-import { CanvasContentLayoutConfig } from "./canvasContentLayoutConfig.js";
 
 /**
  * Base class for all canvas connection segment layout configs
@@ -40,16 +39,13 @@ export abstract class CanvasConnectionSegmentLayoutConfig extends ElementLayoutC
     /**
      * Gets the id of a point registered on the canvas
      *
+     * @param layout the layout engine
      * @param element the current element, its parent must be a point provider (e.g. Canvas)
      * @param pointField the name of the field containing the point
      * @returns the id of the point
      */
-    getContentId(element: LayoutElement, pointField: string): string {
-        const parent = element.parent;
-        if (!parent || !(parent.layoutConfig as CanvasContentLayoutConfig).getContentId) {
-            throw new RuntimeError("CanvasConnectionSegments can only be used as contents of a CanvasConnection");
-        }
+    getContentId(layout: Layout, element: LayoutElement, pointField: string): string {
         const point = element.element.getLocalFieldOrUndefined(pointField)?.value;
-        return (parent.layoutConfig as CanvasContentLayoutConfig).getContentId(element.parent!, point as FullObject);
+        return layout.getElementId(point as FullObject);
     }
 }
