+++ packages/diagram/src/layout/elements/canvas/canvasPointLayoutConfig.ts
@@ -1,6 +1,6 @@
 import { Size } from "@hylimo/diagram-common";
 import { ContentCardinality, LayoutElement, SizeConstraints } from "../../layoutElement.js";
-import { Layout } from "../../layoutEngine.js";
+import { Layout } from "../../engine/layout.js";
 import { CanvasContentLayoutConfig } from "./canvasContentLayoutConfig.js";
 import { nullType } from "@hylimo/core";
